Refactor : component 파일 정리 및 project list 컴포넌트 수정
