+++ src/Screen/MainHome/MainHome.jsx
@@ -1,12 +1,12 @@
 import styles from "./MainHome.module.css";
 
-import MainTitle from "../../Components/MainTitle/MainTitle";
-import MainManageButton from "../../Components/MainManegeButton/MainManageButton";
-import RecommendProject from "../../Components/RecommendProject/RecommendProject";
-import PopularProject from "../../Components/PopularProject/PopularProject";
+import MainTitle from "./MainTitle/MainTitle";
+import MainManageButton from "./MainManegeButton/MainManageButton";
+import RecommendProject from "./RecommendProject/RecommendProject";
+import PopularProject from "./PopularProject/PopularProject";
 
-import Contest from "../../Components/Contest/Contest";
-import News from "../../Components/News/News";
+import Contest from "./Contest/Contest";
+import News from "./News/News";
 
 function MainHome() {
   return (
@@ -17,7 +17,9 @@ function MainHome() {
       <PopularProject />
       <div className={styles.contestnews}>
         <Contest />
-        <div className={styles.news}><News /></div>
+        <div className={styles.news}>
+          <News />
+        </div>
       </div>
       <div className={styles.main}>
         {/*<MyProfile />*/} {/*라우팅 시 outlet이 들어갈 부분*/}
